# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OFPPT Manager API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AlerteNiveau = (typeof AlerteNiveau)[keyof typeof AlerteNiveau];

export const AlerteNiveau = {
  critique: "critique",
  warning: "warning",
  anomalie: "anomalie",
} as const;

export type AlerteEntity = (typeof AlerteEntity)[keyof typeof AlerteEntity];

export const AlerteEntity = {
  stagiaire: "stagiaire",
  groupe: "groupe",
} as const;

export interface Alerte {
  id: string;
  niveau: AlerteNiveau;
  message: string;
  entity: AlerteEntity;
  entityId: string;
  entityLabel: string;
  createdAt: string;
}

export interface ImportLog {
  id: string;
  filename: string;
  type: string;
  nbLignes: number;
  nbErreurs: number;
  warnings: string[];
  dureeMs: number;
  createdAt: string;
}

export interface DashboardData {
  groupesActifs: number;
  tauxMoyen: number;
  tauxTheorique: number;
  alertesCount: number;
  alertesCritiques: number;
  modulesValides: number;
  modulesTotal: number;
  topAlerts: Alerte[];
  recentImports: ImportLog[];
}

export type GroupeAvancementStatut =
  (typeof GroupeAvancementStatut)[keyof typeof GroupeAvancementStatut];

export const GroupeAvancementStatut = {
  en_avance: "en_avance",
  a_jour: "a_jour",
  en_retard: "en_retard",
  non_demarre: "non_demarre",
} as const;

export interface Groupe {
  id: string;
  code: string;
  annee: number;
  mode: string;
  filiereCode: string;
  filiereNom: string;
  statut: string;
  anneeFormation: string;
  tauxReel?: number | null;
  tauxTheorique?: number | null;
  ecart?: number | null;
  avancementStatut: GroupeAvancementStatut;
  nbStagiaires: number;
  createdAt: string;
}

export type CreateGroupeBodyMode =
  (typeof CreateGroupeBodyMode)[keyof typeof CreateGroupeBodyMode];

export const CreateGroupeBodyMode = {
  R_PP: "R_PP",
  R_FF: "R_FF",
  FPA_PP: "FPA_PP",
  FPA_FF: "FPA_FF",
} as const;

export interface CreateGroupeBody {
  code: string;
  mode: CreateGroupeBodyMode;
  filiereCode: string;
  filiereNom: string;
  anneeFormation: string;
  statut?: string;
}

export type GroupeAvancementProjection = {
  dateFin?: string | null;
  retardJours?: number | null;
} | null;

export type ModuleAvancementStatut =
  (typeof ModuleAvancementStatut)[keyof typeof ModuleAvancementStatut];

export const ModuleAvancementStatut = {
  en_avance: "en_avance",
  a_jour: "a_jour",
  en_retard: "en_retard",
  non_demarre: "non_demarre",
  termine: "termine",
  anomalie: "anomalie",
} as const;

export interface ModuleAvancement {
  id: string;
  moduleCode: string;
  moduleIntitule: string;
  mhGlobale: number;
  mhRealise?: number | null;
  tauxReel?: number | null;
  tauxTheorique?: number | null;
  ecart?: number | null;
  statut: ModuleAvancementStatut;
  nbSeancesVal: number;
  nbSeancesEnCours: number;
  sourceFile?: string | null;
  importedAt?: string | null;
}

export interface GroupeAvancement {
  groupeId: string;
  groupeCode: string;
  tauxTheorique?: number | null;
  ecartGlobal?: number | null;
  projection?: GroupeAvancementProjection;
  modules: ModuleAvancement[];
}

export interface Stagiaire {
  id: string;
  cef: string;
  nom: string;
  prenom: string;
  nomComplet: string;
  groupeId: string;
  groupeCode: string;
}

export type NoteModuleEfmStatut =
  (typeof NoteModuleEfmStatut)[keyof typeof NoteModuleEfmStatut];

export const NoteModuleEfmStatut = {
  PRESENT: "PRESENT",
  ABSENT: "ABSENT",
} as const;

export interface NoteModule {
  id: string;
  moduleCode: string;
  moduleIntitule: string;
  cc: number;
  efm: number;
  efmStatut: NoteModuleEfmStatut;
  moyenneOff: number;
  moyenneNorm: number;
  valide: boolean;
  sourceFile: string;
  importedAt: string;
}

export interface StagiaireWithNotes {
  id: string;
  cef: string;
  nom: string;
  prenom: string;
  nomComplet: string;
  groupeId: string;
  groupeCode: string;
  moyenneGenerale?: number | null;
  notes: NoteModule[];
  alertes: Alerte[];
  rang?: number | null;
}

export interface StagiaireDetail {
  stagiaire: Stagiaire;
  notes: NoteModule[];
  alertes: Alerte[];
  moyenneGenerale?: number | null;
}

export interface Module {
  id: string;
  code: string;
  intitule: string;
  mhGlobale: number;
  filiereCode: string;
  niveau: string;
}

export interface ImportResult {
  success: boolean;
  imported: number;
  warnings: string[];
  errors: string[];
  message?: string;
}

export interface CalendrierData {
  id?: string | null;
  typeCalendrier?: string | null;
  totalJours: number;
  joursRealises: number;
  tauxTheorique: number;
  dateReference?: string | null;
  sourceFile?: string | null;
}

export type GetGroupesParams = {
  statut?: string;
};

export type GetStagiairesParams = {
  groupeId?: string;
  search?: string;
};

export type GetAlertesParams = {
  niveau?: GetAlertesNiveau;
  entity?: GetAlertesEntity;
};

export type GetAlertesNiveau =
  (typeof GetAlertesNiveau)[keyof typeof GetAlertesNiveau];

export const GetAlertesNiveau = {
  critique: "critique",
  warning: "warning",
} as const;

export type GetAlertesEntity =
  (typeof GetAlertesEntity)[keyof typeof GetAlertesEntity];

export const GetAlertesEntity = {
  stagiaire: "stagiaire",
  groupe: "groupe",
} as const;

export type ImportFileBodyType =
  (typeof ImportFileBodyType)[keyof typeof ImportFileBodyType];

export const ImportFileBodyType = {
  etat: "etat",
  calendrier: "calendrier",
  pv_efm: "pv_efm",
} as const;

export type ImportFileBody = {
  file: Blob;
  type: ImportFileBodyType;
  groupeId?: string;
};